+++ db/migrations/00010_create_eth_state_accouts_table.sql
@@ -2,7 +2,7 @@
 CREATE TABLE eth.state_accounts (
     header_id             VARCHAR(66) NOT NULL,
     state_path            BYTEA NOT NULL,
-    balance               NUMERIC NOT NULL,
+    balance               BIGINT NOT NULL,
     nonce                 INTEGER NOT NULL,
     code_hash             BYTEA NOT NULL,
     storage_root          VARCHAR(66) NOT NULL,
